refactor: store 객체 생성을 App 내부로 이동
- 해당 App에서만 사용되기 때문에 응집도 높이기 위함
