+++ src/index.ts
@@ -1,7 +1,5 @@
 import App from "./App";
-import RestaurantStore from "./stores/RestaurantStore.js";
 
 window.addEventListener("DOMContentLoaded", () => {
-  const store: RestaurantStore = new RestaurantStore();
-  new App(store);
+  new App();
 });
